move is_gate_applicable_to_cluster func
